pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "traffic-safety"
version = "0.1.0"
description = "Cognitive assessment based driver risk prediction (NIA / KOTSA competition)."
requires-python = ">=3.10"
dependencies = [
    "catboost>=1.2",
    "joblib>=1.3",
    "lightgbm>=4.0",
    "numpy>=1.24",
    "pandas>=2.0",
    "pyarrow>=14.0",
    "pyyaml>=6.0",
    "scikit-learn>=1.3",
    "tqdm>=4.66",
]

[project.scripts]
traffic-safety-train = "traffic_safety.cli:train_cli"
traffic-safety-infer = "traffic_safety.cli:infer_cli"

[tool.setuptools.packages.find]
where = ["src"]
